IR: Working `Array#toString` (#580)
* IR: Working `Array#toString`

The motivating example for this change was getting the
`Array#toString()` method to correctly generate IR, as well as correctly
compiling that IR to both current targets (native and javascript). This
involved implementation of while-loops, loading from memory, and more. I
was also able to fix a bug in the typechecker, related to extracting
generics out of method invocations' `selfVal`.

* Added comment helping to explain the while-condition's 'block'

* introduce 'compound value' rather than 'block' for while condition