IR: Pass the `/compiler/ints` tests
This changeset is using the `/test/compiler/ints.abra` test suite
as a forcing function to drive the implementation of arithmetic
operators and comparisons, as well as introduce floats. This commit
includes unary negate and minus operators, and is working up until the
first encounter with a float (which has not yet been implemented).
That's up next.