IR: simplify array literal generation (#584)
Rather than requiring code generation for the array's `#push` method, we
can insert the items directly into the internal buffer and set the
length to be the known length value. This works for array literals since
all of that information is known at compile-time, and any subsequent
calls to `#push` will still work as expected. But it saves time during
compilation since we don't generate code for the `#push` method unless
it's explicitly called in code (not to mention that inserting directly
into the buffer is more efficient than calling the `#push` method). We
see codegen gains as well as resulting binary size improvements with
this approach because `Array#push` is generic so potential specialized
version are no longer being generated.