IR: default-valued params, `toString` for struct types (#585)
The motivating example here was:
```
val r = range(0, 4)
stdoutWriteln(""+r)
```

This required implementation of functions with default-valued parameters,
in order to call the function, as well as implementation of struct
initializers with default-valued fields. Additionally, this required
emitting structured `toString` output, the core logic for which can be
shared across struct types, enums, and tuples.
This also adds the codegeneration for native code as well as js.