IR: string interpolation
This is working towards getting the `test/compiler/chars.abra` test
suite to pass. One aspect of that was string interpolation, which is now
complete, for both native and js targets. This was relatively
straightforward, and had a nice quick optimization as well for when all
parts of the interpolation are known at compile-time. There's definitely
room for more improvements here, such as constant-folding adjacent
comptime values when the whole thing isn't comptime-known, which would
still be a nice optimization.