IR: Working towards Maps
The motivating example here is to generate IR and compile the following
code to all targets (native/js):
```
val m = { a: 1, b: 2 }
stdoutWriteln(m.toString())
```
Getting this working meant I had to do a lot of side quests, and this
commit is the accumulation of sufficiently many of them. The following
code now works:
```
val m: Map<String, Int> = {}
stdoutWriteln(m.size.toString()) // 0
```
One of the main things that needed changing was the way concrete
generics were resolved when generating function calls. There's now a
notion of "inner" and "outer" generics, meaning that the enqueued
function needs to have its own set of resolved generics (the previous
iteration of the compiler (pre-IR) called this "resolved generic
layers"), along with the parameters and return values. Everything else,
including the arguments (and implicit argument, if it's an instance
method) should use the "outer" concrete generics. This is a lot of words
to express a relatively simple idea.

I also added support for optional targets in if-/while-conditions, as
well as condition binding support for optional values.