IR: Maps work
The motivating example works:
```
val m = { a: 1, "b": 2 }
stdoutWriteln(""+m) // { b: 2, a: 1 }
```

This took a ton of changes in a lot of different places. Basically, in
order to make the `Map#insert` method work, I needed to implement not
only tuples but tuple destructuring. This wasn't a _huge_ undertaking in
itself, but it did mean that I needed to add toString and eq support for
them as well (and eq support in general, since that was a thing that
didn't yet exist for any non-primitive type). I also needed to implement
`toString`/`eq`/`hash` substitution functionality for when those methods
are called on a generic type.
Other than that there were indexing changes, assignment changes, changes
to the way in which binding patterns are evaluated, and also to the way
while-loops work. The previous mechanism of doing while-loops was
insufficient when considering the need to destructure the value as well
as compare it to its zero/null value.
Not to mention many bugfixes along the way, like for bitwise and/or only
considering the `result` of their block and not evaluating the whole
body. Or also how adding an int and a byte value would result in a F64
(??).
Additionally, in the js compilation target, I changed from having
strings be an array of characters (really, an array of strings since
characters don't exist in javascript) to instead having a string be
backed by a `Uint8Array` buffer. This makes it possible to do all of the
manual memcpy and insert logic that exists throughout the toString
implementations (like `Array#toString` for example).