Special for-loop form
Originally the `range` function was meant to be the primary mechanism
for iterating over a range of ints. However, this is actually a fairly
heavyweight operation for what it is (calling `range`, which
instantiates an object, then successively calling `next()` on that
object until it returns `None`; versus simply incrementing a number in a
local variable). This is the first pass of the change; a subsequent pass
will update the standard library to use this style of for-loop. The
`range` method itself will remain since it is still useful (for example,
it allows for specifying a `stepBy` value in cases where you don't want
to increment by 1 each time).