For-loop cleanup (and tests) (#591)
* For-loop cleanup (and tests)

Replace nearly all instances of `range()` in for-loops with the
simplified special ranged for-loop form. This is cleaner and also much
more efficient, and almost entirely removes the need for the `range`
function. It also greatly simplifies IR generation for for-loops, since
it does not require any complex iterator logic to be implemented in
order to get started.
Also, this removes all of the anonymous lambdas that had been created
during the compiler tests - instead it replaced it with `if true {...}`
construction, which is effectively just an isolated scope. In the
future, this may be replaced with a construction like `do { ... }`, but
more likely it'll be replaced with some kind of built-in testing
capabilities.

* one tiny thing