IR: Working towards Maps (#589)
* IR: Working towards Maps

The motivating example here is to generate IR and compile the following
code to all targets (native/js):
```
val m = { a: 1, b: 2 }
stdoutWriteln(m.toString())
```
Getting this working meant I had to do a lot of side quests, and this
commit is the accumulation of sufficiently many of them. The following
code now works:
```
val m: Map<String, Int> = {}
stdoutWriteln(m.size.toString()) // 0
```
One of the main things that needed changing was the way concrete
generics were resolved when generating function calls. There's now a
notion of "inner" and "outer" generics, meaning that the enqueued
function needs to have its own set of resolved generics (the previous
iteration of the compiler (pre-IR) called this "resolved generic
layers"), along with the parameters and return values. Everything else,
including the arguments (and implicit argument, if it's an instance
method) should use the "outer" concrete generics. This is a lot of words
to express a relatively simple idea.

I also added support for optional targets in if-/while-conditions, as
well as condition binding support for optional values.

* IR: Maps work

The motivating example works:
```
val m = { a: 1, "b": 2 }
stdoutWriteln(""+m) // { b: 2, a: 1 }
```

This took a ton of changes in a lot of different places. Basically, in
order to make the `Map#insert` method work, I needed to implement not
only tuples but tuple destructuring. This wasn't a _huge_ undertaking in
itself, but it did mean that I needed to add toString and eq support for
them as well (and eq support in general, since that was a thing that
didn't yet exist for any non-primitive type). I also needed to implement
`toString`/`eq`/`hash` substitution functionality for when those methods
are called on a generic type.
Other than that there were indexing changes, assignment changes, changes
to the way in which binding patterns are evaluated, and also to the way
while-loops work. The previous mechanism of doing while-loops was
insufficient when considering the need to destructure the value as well
as compare it to its zero/null value.
Not to mention many bugfixes along the way, like for bitwise and/or only
considering the `result` of their block and not evaluating the whole
body. Or also how adding an int and a byte value would result in a F64
(??).
Additionally, in the js compilation target, I changed from having
strings be an array of characters (really, an array of strings since
characters don't exist in javascript) to instead having a string be
backed by a `Uint8Array` buffer. This makes it possible to do all of the
manual memcpy and insert logic that exists throughout the toString
implementations (like `Array#toString` for example).

* minor cleanup and test fixes

* IR: Generating IR for (ranged) for-loops

Only for ranged for-loops for now, just to get Maps working.

* For-loops for arrays and maps

(And Sets too, but Set literals aren't handled yet in IR generation).
This also fixes Option-type negation.