CLI: Building the cli natively, rather than sh
In order to accomplish this, there were quite a few things that had to
be done. I needed some way of executing external programs from abra (to
call out to `qbe` and `clang`), and to do that I needed to do a lot of
groudwork with forking, creating pipes to redirect and capture io, etc.
There are a _ton_ of improvements to be made here, but I wanted to take
this first initial step (which does work! it supports the `build`
subcommand) and commit it because it represented a lot of work.