IR: Getting `chars` test to pass
The goal was to get the `test/compiler/chars.abra` test suite to pass,
which involved implementing a number of IR features as well as some
bugfixes. The IR now supports (non-capturing) function values as well as
rudimentary `try` expressions; these features are also carried through
to the native and js compilation targets.
This also fixes a bug where bitwise AND-ing a Byte and an Int would
result in a Bool type (??), and it also fixes a typo where shift-left
and shift-right resulted in a Bool type (whoops).
I also implemented the coalesce (?:) operator (which I actually think
I'll one day deprecate in favor of `try`-`else` expression, since it's
identical in function), as well as basic `hash` method support for
primitive types (Int, Bool, Float, and Char).