VM for IR
This implements another target for the IR: execution in a VM. This is
ultimately what will power compile-time execution of functions
(macros?). This reaches parity with the existing coverage of the IR, and
will continue to improve as the IR coverage improves (it has a test
suite which runs against the tests that the native and js targets do).