IR: Functions as values (including closures) (#596)
* IR: Functions as values (including closures)

This is a work in progress

* Function values work now

This completes the initial work to have lambdas, function values, and
closures. There's still some additional work remaining to create
function values from _function_ references (not just lambdas), and
especially from method references (eg. `val f = [1, 2, 3].isEmpty`).
But this is a great start, and it works in all IR targets (including the
newly-added VM target!).
One notable thing is that the original `test/compiler/functions` test
suite wasn't actually testing any closures; the would-be closures were
top-level functions which were closing over top-level variables, and
that actually doesn't count as a closure. Top-level variables are
treated as globals if they're referenced in any functions, so no closure
environment or function value is created. So, I created
`test/compiler/functions'` as a way of testing the intended
functionality of closures, but there are some limitations. For example,
function declarations currently cannot be contained within other
functions. So, the only way to create a closure at the moment is with a
lambda. Creating lambdas introduces some additional concerns, such as
default-valued parameters. I've captured all of this in the `functions'`
test suite, which passes against all IR targets as well as the original
implementation of the compiler.