IR VM cleanup (#599)
The initial implementation of the VM which interprets generated IR was
to emulate the low-level operations 1-1 (including raw pointer
operations). This proved to be fairly difficult because it required
maintaining two separate object structures: there was the VmValue type
which represented known locals/captures/etc, and then there was the
actual in-memory representation of the data. Converting between the two
was pretty ungainly and error-prone. This introduces a `VmValue.Instance`
variant, which lifts all of the loadfield/storefield operations out of
the low-level memory space and into emulated space. Ultimately it is
most likely less performant, but it's a good tradeoff for readability
(at the moment).