Macros: proof-of-concept (#600)
* Macros: proof-of-concept

One of the main motivations for introducing IR (aside from the ability
to perform optimizations, and other motivations) was to compile to many
targets, including js but especially to a VM. Using this VM, we can
write special functions (decorated with `@macro`) which will be compiled
to IR and executed _at compile-time_ (specifically during typechecking).
While I'm still working out the finer details of macros, this represents
a functional proof-of-concept.

* Introduce `meta` module

Add some typechecking and introduce the `meta` std module