Macros: error reporting (#602)
* Macros: error reporting

Surface errors when defining/invoking macro functions. If macro
functions have parameters of incorrect type (as of right now, the only
valid param type for macro functions is `meta.Expr`) or incorrect return
type (`meta.InjectedCode`), then a typechecker error will be raised for
that function declaration.
If a macro function has any errors in its body (or module, really,
since at the moment it's difficult/impossible to determine whether the
function itself has any errors in it), then we are not guaranteed to be
able to run the macro function at compile-time, and raise an error at
the macro invocation site (the underlying errors will also be surfaced
in the error report). If the code _produced_ by the macro function at
compile-time fails to be tokenized, parsed, or typechecked, then a
similar error will be raised at the macro invocation site.

* fixing tests