Changing typechecking order
Previously, the typechecking algorithm operated in a depth-first manner;
given a module, it'd fully typecheck any of its dependencies
(recursively), and then fully typecheck the entrypoint module. This
meant that all top-level declarations, _as well as_ the bodies of those
functions/methods, would be typechecked at the same time. This made
certain things simply impossible; for example, the `prelude` module
(upon which every other module implicitly relies) cannot itself pull in
any dependencies which make use of any of the types/methods defined
therein. As a concrete example, I've wanted to add a `toString` method
to the `intrinsics.Pointer` type for a long time, with a body of:
```
func toString(self): String = "Pointer(${self.address().hex()})"
```
but I could not, because at the time of typechecking the intrinsics
module (which `prelude` depends on), the entirety of `prelude` hadn't
yet been typechecked, and the `Int#hex` method didn't exist yet. Now
that all declarations are discovered prior to typechecking the
definitions of those declarations, this is possible.

Additionally (and perhaps most desiredly) this allows for macro
functions to be defined and used within the `prelude` module. This means
that `println`, `debug`, `assert`, and others can be easily surfaced
from `prelude`, but it also means that they can be used from within the
module as well (e.g. creating a `Array#expect(Int)` method which does an
`assert` to check the bounds of the array).

This also adds potential for parallel processing of modules'
declarations (although that's a _long long long_ ways off), and has also
acted as a forcing function for some cleanup that's been hanging around
for a while. Note that there are still some things that are broken -
namely, exported variables (exported types/enums work though). I wanted
to get this commit in first though, because I think that'll possibly be
a tricky thing to get right. However, if I do then it might simplify
other things as well (such as the weird way I've been handling functions
with default param values which reference declarations in the module).