Coalesce the default params logic and binding decl logic
Combine these two into a single notion - there is a way to temporarily
capture any errors encountered when typechecking an expression, and
there is also a way to capture certain special events that happen during
the typechecking of an expression (namely when a function call happens,
or a function variable alias is referenced). If there are errors
captured, or any of those special events occur, then we know that we
need to re-attempt typechecking the expression in a later phase. This
holds true both for typechecking of top-level variables but also for
functions with default-valued parameters.

In addition to this change, I also simplified a lot of the
default-valued parameter logic. Now, params' variables are added to the
local scope even if they're incomplete; this allows for fewer redos in
the later phase.