Changing typechecking order (#603)
* Changing typechecking order

Previously, the typechecking algorithm operated in a depth-first manner;
given a module, it'd fully typecheck any of its dependencies
(recursively), and then fully typecheck the entrypoint module. This
meant that all top-level declarations, _as well as_ the bodies of those
functions/methods, would be typechecked at the same time. This made
certain things simply impossible; for example, the `prelude` module
(upon which every other module implicitly relies) cannot itself pull in
any dependencies which make use of any of the types/methods defined
therein. As a concrete example, I've wanted to add a `toString` method
to the `intrinsics.Pointer` type for a long time, with a body of:
```
func toString(self): String = "Pointer(${self.address().hex()})"
```
but I could not, because at the time of typechecking the intrinsics
module (which `prelude` depends on), the entirety of `prelude` hadn't
yet been typechecked, and the `Int#hex` method didn't exist yet. Now
that all declarations are discovered prior to typechecking the
definitions of those declarations, this is possible.

Additionally (and perhaps most desiredly) this allows for macro
functions to be defined and used within the `prelude` module. This means
that `println`, `debug`, `assert`, and others can be easily surfaced
from `prelude`, but it also means that they can be used from within the
module as well (e.g. creating a `Array#expect(Int)` method which does an
`assert` to check the bounds of the array).

This also adds potential for parallel processing of modules'
declarations (although that's a _long long long_ ways off), and has also
acted as a forcing function for some cleanup that's been hanging around
for a while. Note that there are still some things that are broken -
namely, exported variables (exported types/enums work though). I wanted
to get this commit in first though, because I think that'll possibly be
a tricky thing to get right. However, if I do then it might simplify
other things as well (such as the weird way I've been handling functions
with default param values which reference declarations in the module).

* Change top-level binding typechecking order

Typechecking is now done in two discrete passes - the first pass
discovers all of the declarations (recursively throughout all imported
modules); the second pass typechecks all of the definitions (e.g. struct
fields, enum variants, and instance/static methods). In the first pass,
we need to be able to discover the top-level bindings (declared with
val/var) and handle them in a similar way. The main complexity here is
knowing if/when to defer typechecking of the assignment expression,
and when it potentially needs to be re-checked later on when additional
information is available. There's certainly improvements to be made, but
everything appears to be working after this great shift.

* Name lambdas based on location rather than autoincrementing idx

* Coalesce the default params logic and binding decl logic

Combine these two into a single notion - there is a way to temporarily
capture any errors encountered when typechecking an expression, and
there is also a way to capture certain special events that happen during
the typechecking of an expression (namely when a function call happens,
or a function variable alias is referenced). If there are errors
captured, or any of those special events occur, then we know that we
need to re-attempt typechecking the expression in a later phase. This
holds true both for typechecking of top-level variables but also for
functions with default-valued parameters.

In addition to this change, I also simplified a lot of the
default-valued parameter logic. Now, params' variables are added to the
local scope even if they're incomplete; this allows for fewer redos in
the later phase.

* removing outdated comment