Bugfix: ensuring scope is restored
When discovering declarations in the first phase of typechecking a
module, it's possible the initialization expression of a binding may
result in a failure. This could leave the `currentScope` in a weird
state after typechecking the initialization expression, and since the
language doesn't have any `defer` mechanism (and likely won't, at least
not for the foreseeable future), we can just manually restore the scope
afterwards.