Add macros to prelude (#611)
* Add macros to prelude

This adds the `println` and `debug` macro to the `prelude` module, the
culmination of the big refactoring as well as all of the macro work
(which includes the IR generation, as well as the VM). The next PR will
see the `stdoutWrite`/`stdoutWriteln` be replaced by `print`/`println`
macros in the test suites, followed by the removal of those functions.

* adding more stringify support

* Update compiler tests

Replace instances of `stdoutWriteln` with the `println` macro in the
compiler test suite