IR: Sets and Maps (#613)
This adds support for Set literals into the IR compiler. Also, this
enables the set and map test suites for IR generation (for all targets).
In doing this, I discovered errors with the way I had been handling
concrete generics for for-loop bindings. This required a _lot_ of work
to dig into and fix, but luckily the solution was simple (albeit clunky,
since it involves an additional traversal of the for-loop binding
pattern).