IR: Enabling strings and arrays test suites (#614)
In order to enable these and have them pass against all IR targets,
there were quite a few things I needed to do. Namely, I needed to
finally properly represent `Float?` values (which I represented with the
help of IEEE 754 Quiet NaN values). Also, there's a pre-existing issue
with the `println` implementation (rather, the implementation of macros
in general) as it relates to lambdas. Lambdas' names are derived based
on their position in the source code (ordinarily this acts as a stable
identifier); however when the generated code gets emitted from the macro
runner and parsed into a module, the generated lambda names conflict.
For now, I refactored the tests to avoid having lambdas in the `println`
call.