Remove `Type` match case kind
This was a carryover from a previous implementation of the language,
where a type name could be used as part of a match case, for example:
```
val arr = [1, 2, 3]
val x = match arr[0] { Int i => i + 1, None => 123 }
```
This however proved to be not very useful since other language
constructs already exist to support this use-case, let alone the match
itself:
```
match arr[0] { None => 123, else i => i + 1 }
```
This idea also was part of an original union-types idea from way back in
the early days of the language, where it was possible to create types
such as:
```
val a: Int | Float = ...
match a {
  Int i => println(i)
  Float f => println(f)
}
```
However, this idea has long since been retired, and will not be
reintroduced as that's not the direction I want to take the language in.

Removing this match case kind simplifies the implementaiton of compiling
match statements/expressions, and is ultimately a step towards
generating simpler IR (which is why I'm doing it now).