Handling match statements/expressions
This implements basic handling of match statements/expressions,
including handling None cases, literal cases, and else cases. This also
includes an optimization for catching potential None values (for
Option-type subject values) early, and unwrapping for subsequent cases
if non-None (this is better than checking for None in each case, since
we can assume if control flow makes it past the None case, if it exists,
then the value is non-None; and if no None case exists then the
else-case can catch it).