Support matching on enum variant cases
This adds case testing for enum variants, both constant and container
(for containers, it also adds the ability to destructure the fields into
variables which can be used in the block). This also fixes a bug in the
literal case test, because we should use the equality logic rather than
always using the Eq operation (strings should not be compared by
reference).