IR: Enabling `compiler/optionals.abra` test suite (#615)
* IR: Enabling `compiler/optionals.abra` test suite

This enables the test suite for all IR targets. In order to accomplish
this, I first needed to implement optional-safe accessing and
invocation, as well as the flattened representation of None for `Bool?`
(which is just `2`, since Bools are 64-bit integers).

* Add support for enums (just to enable optional suite)

The `compiler/optionals.abra` test suite includes operations with enums,
which hadn't been implemented yet. This implements nearly all features
of enums themselves, but not `match` statements/expressions, which are
required in order to unlock the `compiler/enums.abra` suite. So that'll
be up next!

* Remove `Type` match case kind

This was a carryover from a previous implementation of the language,
where a type name could be used as part of a match case, for example:
```
val arr = [1, 2, 3]
val x = match arr[0] { Int i => i + 1, None => 123 }
```
This however proved to be not very useful since other language
constructs already exist to support this use-case, let alone the match
itself:
```
match arr[0] { None => 123, else i => i + 1 }
```
This idea also was part of an original union-types idea from way back in
the early days of the language, where it was possible to create types
such as:
```
val a: Int | Float = ...
match a {
  Int i => println(i)
  Float f => println(f)
}
```
However, this idea has long since been retired, and will not be
reintroduced as that's not the direction I want to take the language in.

Removing this match case kind simplifies the implementaiton of compiling
match statements/expressions, and is ultimately a step towards
generating simpler IR (which is why I'm doing it now).

* Handling match statements/expressions

This implements basic handling of match statements/expressions,
including handling None cases, literal cases, and else cases. This also
includes an optimization for catching potential None values (for
Option-type subject values) early, and unwrapping for subsequent cases
if non-None (this is better than checking for None in each case, since
we can assume if control flow makes it past the None case, if it exists,
then the value is non-None; and if no None case exists then the
else-case can catch it).

* Support matching on enum variant cases

This adds case testing for enum variants, both constant and container
(for containers, it also adds the ability to destructure the fields into
variables which can be used in the block). This also fixes a bug in the
literal case test, because we should use the equality logic rather than
always using the Eq operation (strings should not be compared by
reference).