Working towards method-reference values
These will be implemented as closures which capture the value of the
implicit argument (`self`). For now, let's abstract out the notion of
invoking a function value, in the native compilation target, since it's
already fairly complex logic and it'll be nicer to augment that logic
once it's tucked away into a function.