Uncomment the `compiler/process.abra` suite (#620)
There were a handful of builtins that needed to be implemented across
the IR compilation targets in order to unlock this test suite.
Additionally, there were a handful of bugs that I caught along the way.
There were some bugs that I fixed along the way, and I also realized
that I hadn't updated the html-based js runtime harness in a while.