Type-hole filling improvements (#623)
* Type-hole filling improvements

Some improvements to the typechecker and how it handles filling in type
holes during inference. There's plenty of more work to be done here (see

* Filling type holes at AST level

There are plenty of cases where updating incomplete type information
requires types to be propagated throughout the AST. This addresses some
(most?) of the issues with incomplete `resolvedGenerics` when compiling
invocations of generic functions.